Use a long type for entity ids to expand the potential pool of ids

+++ src/main/raptor/engine/game/entity/BasicEntity.java
@@ -5,13 +5,13 @@ import raptor.engine.logical.collision.api.ICollisionShape;
 import raptor.engine.util.geometry.Point;
 
 public abstract class BasicEntity implements IEntity {
-	private final int id;
+	private final long id;
 	private final IEventSource eventSource;
 	private final ICollisionShape collision;
 
 	private final Point position;
 
-	public BasicEntity(final int id, final IEventSource eventSource, final ICollisionShape collision) {
+	public BasicEntity(final long id, final IEventSource eventSource, final ICollisionShape collision) {
 		this.id = id;
 		this.eventSource = eventSource;
 		this.collision = collision;
@@ -20,7 +20,7 @@ public abstract class BasicEntity implements IEntity {
 	}
 
 	@Override
-	public int getId() {
+	public long getId() {
 		return id;
 	}
